##### Note: GQview CVS on sourceforge is not always up to date, please use ##### ##### an offical release when making enhancements and translation updates. #####
Thu Mar  3 14:24:58 2005  John Ellis  <johne@verizon.net>

        * filelist.c, info.c, preferences.c, utilops.c: Use doubles instead of
        floats whenever possible.
        * ui_utildlg.c (generic_dialog_add_message): Enable line wrap for
        message body so specifying newlines is no longer needed.
        * pan-view.c: Display info dialog when thumbnail caching settings are
        not optimal.

+++ src/filelist.c
@@ -514,16 +514,16 @@ gchar *text_from_size_abrev(gint64 size)
 		}
 	if (size < (gint64)1048576)
 		{
-		return g_strdup_printf(_("%.1f K"), (gfloat)size / 1024.0);
+		return g_strdup_printf(_("%.1f K"), (double)size / 1024.0);
 		}
 	if (size < (gint64)1073741824)
 		{
-		return g_strdup_printf(_("%.1f MB"), (gfloat)size / 1048576.0);
+		return g_strdup_printf(_("%.1f MB"), (double)size / 1048576.0);
 		}
 
-	/* to avoid overflowing the float, do division in two steps */
-	size /= 1048576.0;
-	return g_strdup_printf(_("%.1f GB"), (gfloat)size / 1024.0);
+	/* to avoid overflowing the double, do division in two steps */
+	size /= 1048576;
+	return g_strdup_printf(_("%.1f GB"), (double)size / 1024.0);
 }
 
 /* note: returned string is valid until next call to text_from_time() */
